# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AUTH API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Posts",
                "description": "Posts list",
                "operationId": "5a796801ed2e8bd6a779c472ca9bc119",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "A list of posts.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": ""
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "created_at": {
                                                        "type": "number",
                                                        "example": ""
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 12,
                                                    "title": "Первый пост",
                                                    "content": "Текст поста",
                                                    "created_at": "2023-06-25 00:58:50"
                                                },
                                                {
                                                    "id": 13,
                                                    "title": "Первый пост 2",
                                                    "content": "Текст поста",
                                                    "created_at": "2023-06-25 00:59:47"
                                                }
                                            ]
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "https://ovva.local/api/v1/posts?page=1"
                                        },
                                        "from": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "number",
                                            "example": 3
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "https://ovva.local/api/v1/posts?page=3"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": ""
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "url": "https://ovva.local/api/v1/posts?page=1",
                                                    "label": "« Previous",
                                                    "active": false
                                                },
                                                {
                                                    "url": "https://ovva.local/api/v1/posts?page=1",
                                                    "label": "1",
                                                    "active": false
                                                },
                                                {
                                                    "url": "https://ovva.local/api/v1/posts?page=2",
                                                    "label": "2",
                                                    "active": true
                                                },
                                                {
                                                    "url": "https://ovva.local/api/v1/posts?page=3",
                                                    "label": "3",
                                                    "active": false
                                                },
                                                {
                                                    "url": "https://ovva.local/api/v1/posts?page=3",
                                                    "label": "Next »",
                                                    "active": false
                                                }
                                            ]
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "https://ovva.local/api/v1/posts?page=3"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "https://ovva.local/api/v1/posts"
                                        },
                                        "per_page": {
                                            "type": "number",
                                            "example": 2
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "https://ovva.local/api/v1/posts?page=1"
                                        },
                                        "to": {
                                            "type": "number",
                                            "example": 4
                                        },
                                        "total": {
                                            "type": "number",
                                            "example": 6
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Creat post",
                "description": "Creat post",
                "operationId": "f76a4167e19d1f4ee1c3dd53903058cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Первый Пост"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Описание первого поста"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Первый Пост"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Описание первого поста"
                                        },
                                        "id": {
                                            "type": "number",
                                            "example": 16
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        },
                                        "Description": {
                                            "type": "string",
                                            "example": "Missing or Invalid Access Token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Поле 'title' обязательное!"
                                                    }
                                                },
                                                "content": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Поле 'content' обязательное!"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}